updates to my readme and iterating the halo.
I continued to update my readme document with where I was up to in the coding process, and describing my troubles and what I was trying to achieve. I really wanted a glowing halo behind my cat, and this was taking a while to get right – I was testing many colours and the sizing of the ellipses.