Update to readme and fixing my slow loading problem
I've updated my readme with my latest tests of the AI and iterations of the design of my code. I also got help to fix the slowness of loading in the pixels on top of my cat (the third layer). With help I also fixed the issue I was having with the transparency not working.