Error fixed in index.html for social links
I was struggling to work out why my link to my project wasn't working (and gave me a 404 error) when I tested it. It turns out it was as simple as missing out a hyphen. So, all fixed now! Also I made some adjustments to the background of my projects – giving a blurry/faded effect using greyscale and ellipses with various opacity.